+++ Sources/Plugins/HtmlMinify.php
@@ -148,7 +148,7 @@ function eve_html_compression_start(): void {
     ob_start('\\WordPress\\Themes\\EveOnline\\Plugins\\eve_html_compression_finish');
 }
 
-$themeOptions = get_option('eve_theme_options', ThemeHelper::getThemeDefaultOptions());
+$themeOptions = get_option('eve_theme_options', ThemeHelper::getInstance()->getThemeDefaultOptions());
 
 if (!empty($themeOptions['minify_html_output']['yes'])) {
     add_action('get_header', '\\WordPress\\Themes\\EveOnline\\Plugins\\eve_html_compression_start');