+++ page.php
@@ -33,7 +33,7 @@ get_header();
 
     <div class="container main">
         <?php
-        $breadcrumbNavigation = NavigationHelper::getBreadcrumbNavigation();
+        $breadcrumbNavigation = NavigationHelper::getInstance()->getBreadcrumbNavigation();
 
         if (!empty($breadcrumbNavigation)) {
             ?>
@@ -57,7 +57,7 @@ get_header();
                 ?>
                 <!--<div class="row main-content">-->
                 <div class="main-content clearfix">
-                    <div class="<?php echo PostHelper::getMainContentColClasses(); ?> content-wrapper">
+                    <div class="<?php echo PostHelper::getInstance()->getMainContentColClasses(); ?> content-wrapper">
                         <div class="content content-inner content-page">
                             <header>
                                 <?php
@@ -103,15 +103,15 @@ get_header();
             }
         }
 
-        if (ThemeHelper::hasSidebar('sidebar-page') || ThemeHelper::hasSidebar('sidebar-general')) {
+        if (ThemeHelper::getInstance()->hasSidebar('sidebar-page') || ThemeHelper::getInstance()->hasSidebar('sidebar-general')) {
             ?>
             <div class="col-lg-3 col-md-3 col-sm-3 col-3 sidebar-wrapper">
                 <?php
-                if (ThemeHelper::hasSidebar('sidebar-general')) {
+                if (ThemeHelper::getInstance()->hasSidebar('sidebar-general')) {
                     get_sidebar('general');
                 }
 
-                if (ThemeHelper::hasSidebar('sidebar-page')) {
+                if (ThemeHelper::getInstance()->hasSidebar('sidebar-page')) {
                     get_sidebar('page');
                 }
                 ?>